+++ src/lib.rs
@@ -4,6 +4,7 @@
 #![allow(dead_code)]
 #![allow(unused_imports)]
 #![feature(alloc_error_handler)]
+#![feature(const_mut_refs)]
 #![feature(custom_test_frameworks)]
 #![feature(abi_x86_interrupt)]
 #![test_runner(crate::testexec)]